Updated add Non University Location page to allow the user to enter a name that is up to 20 characters long and tied the format of the name to an applicatoin property.
The user can set the property tmtbl.nonUniversityLocation.pattern to a regular expression in order to specify a specific pattern the Non University Location names must conform to.  If tmtbl.nonUniversityLocation.pattern is not set the name is checked to make sure it only contains the standard set valid characters.

tmtbl.nonUniversityLocation.patternInfo can be used to provide an error message when the name does not match the regular expression.


Former-commit-id: 4bc3d52f09cad0a246192711756dd979eeb2093a