Updated default session selection to default based on user role when there is only one academic initiative.
No Role - the default will be the session the current date falls within, if the current date does not fall with a session it will select the first future session, if there is no future session it will select the most recent past session.

Exam Manager - the default will be the session the current date falls within if final exams have not been published, if final exams have been published it will select the first future session, if the current date does not fall with a session it will select the first future session, if there is no future session it will select the most recent past session.

Event Manager - the default will be the session the current date falls within, if the current date does not fall with a session it will select the first future session, if there is no future session it will select the most recent past session.

View All - the default will be the session the current date falls within, if the current date does not fall with a session it will select the first future session, if there is no future session it will select the most recent past session.

Department Schedule Manager - the first future session, if there is no future session it will select the current session, if there is not current session it will select the most recent past session.

Administrator - the default will be the session the current date falls within, if the current date does not fall with a session it will select the first future session, if there is no future session it will select the most recent past session.


Former-commit-id: 5985b4a70735e5638334b32b641e3922176f0983