Database changes (room ratio, fixed units)
Unfortunately the fields class_.room_ratio and course_credit_unit_config.fixed_units were defined as NUMBER (without any parameters) in Oracle, and migrated as DECIMAL(22,0) to MySQL. This change changes them to DOUBLE in MySQL. The fix is only needed when MySQL is used, and the database was initially created using schema.sql from UniTime 3.2 (as UniTime 3.1 fixed the room ratio in change 37).

This change also corrects the automatic updates 35-37 which did not have schema name attached to the table name.


Former-commit-id: 8ac761cdad1a0eff2ee139cc09fe6f8a8310fa43