Online Student Sectioning -- merge student sectioning changes from the development (dev_sectioning branch) to the trunk
Although the online student sectioning still does not have all the desired functionality, there are a lot of improvements to the Student Scheduling Assistant which the 3.2 version can benefit from. Since the related code has changed a lot it is impossible to cherry pick just the scheduling assistant changes and it does not make any sense to keep two separate versions (especially since the online sectioning functions are only enabled when the session status allows for that). The development will still continue to be done on the dev_sectioning branch.


Former-commit-id: 5cc2039cff4547effb1cdbda598c9f4e7b83f615