Point In Time Data XML Import
Added ability to Import a Point In Time Data XML Export
- The export can be taken from a different instance of UniTime.
- If the point in time data being imported matches an existing session
within UniTime the data is imported into that session
- If the point in time data being imported does not match an existing
session within UniTime a new session corresponding to the snapshot will
be created that contains only the data from the snapshot.
- The expected way to populate Point In Time Data within UniTime is to
do an XML export of Point In Time Data from UniTime and then import the
data back into UniTime.


Former-commit-id: fc619485aacda29bb5698a5e5440eb4f6c7ad447