Date Pattern Change Refinement
Refined the message for date pattern changes and updated the release
notes.

An issue with the quoting in the date pattern change message has been
corrected.

If a class has both a Date Pattern change and a Meeting Conflict, only
one warning icon is displayed. The mouseover for the icon contains both
warnings.


Former-commit-id: 1705a714e8b99a91abf6cf91fc5cda7936ed2fde