Assign Instructors Page GWT Rewrite
The Assign Instructors Page has been rewritten in GWT and has had support for the funding department column added.

The new page is used by default if it is desired to use the old version of the page it can be used by setting the "unitime.legacy.course.assign_instructors" property to "true".


Former-commit-id: bf29d6400413c9d743df34f9ceddff2aca11fabb