Update Change Logging
When using a session context, if a subject area or department is defined use the subject area or department to get the academic session rather than the user context.  This ensures more accurate logging of the change to the correct academic session when a change is made via an API call.


Former-commit-id: 511b39189827be13fefbd7b1f170632da8579c65